+++ packages/at_client_flutter/lib/src/keychain/keychain_store.dart
@@ -1,4 +1,4 @@
-import 'package:package_info_plus/package_info_plus.dart';
+part of 'keychain_storage.dart';
 
 const _atKeysStoreName = '@atsigns';
 const _enrollmentStoreName = '@enrollment';
@@ -18,8 +18,14 @@ class AtKeysStore extends KeychainStore {
   const AtKeysStore();
   static Future<String> getName() async {
     String packageName = await getPackageName();
-    return '${_atKeysStoreName}_$packageName';
+    return '$_atKeysStoreName:$packageName';
   }
+}
+
+// for < 1.1.6 store name (used `_` instead of `:`).
+Future<String> _getImproperAtKeysStoreName() async {
+  String packageName = await getPackageName();
+  return '${_atKeysStoreName}_$packageName';
 }
 
 class EnrollmentStore extends KeychainStore {